backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
onnxruntime==1.24.1
Pillow>=10.0.0
numpy>=1.26.0
huggingface_hub==0.23.0